fix(bd-6pi): split config aggregator into per-condition halves
The neutral plugin/config/index.ts re-exported BOTH
createHandlerOptions.client AND createHandlerOptions.server. Anything
importing from "vite-plugin-react-server/config" (e.g. bidoof's
vite.config.ts uses getCondition from that subpath) statically links
the .client side, which transitively reaches vendor.client.js, which
calls projectRequire("react-dom/server") at module init. Under
--conditions react-server that resolves to react-dom 19.x's
server.react-server.js whose first line throws by design — crashing
both dev and build on every consumer that runs under react-server.

Split into plugin/config/index.server.ts and plugin/config/index.client.ts,
each re-exporting the side-correct createHandlerOptions and only
condition-neutral utilities. Make ./config a conditional package
export and update internal callers to depend on specific files
(defaults, autoDiscover/resolveAutoDiscover) rather than the
aggregator.

Root cause for the bidoof regression bisected to PR #27 in bd-6pi.
